Succesfully ran application after refactor into Blueprint and app factory model.
