+++ .flaskenv
@@ -1,1 +1,2 @@
-FLASK_APP=app.py
+FLASK_APP=app
+FLASK_ENV=development
