+++ app/models/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+blueprint = Blueprint('models', __name__)
+
+from app.models import entities
