+++ app/static/script.js
@@ -1,5 +1,5 @@
 const navBar = document.getElementById('topnav');
-var userName = "username";
+let userName = "username";
 
 
 displayNavBar();
@@ -17,7 +17,7 @@ function toggleMenu() {
 document.addEventListener('click', function (event) {
     submenu = document.getElementById("submenu-container");
     userButton = document.getElementById("usernameButton");
-    var isClickInside = (submenu.contains(event.target) || userButton.contains(event.target));
+    let isClickInside = (submenu.contains(event.target) || userButton.contains(event.target));
     if (!isClickInside && submenu.classList.contains("open-menu")) {
         submenu.classList.remove("open-menu");
     }
@@ -32,7 +32,7 @@ function toggleContactInfo() {
 }
 
 document.addEventListener('click', function (event) {
-    var isClickInside = (contactContainer.contains(event.target) || contactButton.contains(event.target));
+    let isClickInside = (contactContainer.contains(event.target) || contactButton.contains(event.target));
     if (!isClickInside && contactContainer.classList.contains("open-menu")) {
         contactContainer.classList.remove("open-menu");
     }
@@ -51,7 +51,7 @@ function toggleLoginMenu() {
 document.addEventListener('click', function (event) {
     loginmenu = document.getElementById("loginmenu-container");
     loginButton = document.getElementById("login-button");
-    var isClickInside = (loginmenu.contains(event.target) || loginButton.contains(event.target));
+    let isClickInside = (loginmenu.contains(event.target) || loginButton.contains(event.target));
     if (!isClickInside && loginmenu.classList.contains("open-menu")) {
         loginmenu.classList.remove("open-menu");
     }
@@ -60,16 +60,17 @@ document.addEventListener('click', function (event) {
 
 //Used to create the navigation bar at the top of the page. differs depending on whether a user is logged in or not.
 function displayNavBar() {
-    if (userName != "") {
-        navBar.innerHTML = "<a href='/'><img src='/static/images/banner.png' alt='Team 1 Travel Planner banner'"
-            + "width='961' height='51'></a> <button class='button signup' id='usernameButton' onclick='toggleMenu();'>" + userName + "</button>"
+    if (userName !== "") {
+        navBar.innerHTML = "<a href='/static'><img src='/static/images/banner.png' alt='Team 1 Travel Planner banner'"
+            + " width='961' height='51'></a> <button class='button signup' id='usernameButton'" +
+            " onclick='toggleMenu();'>" + userName + "</button>"
             + "<div id='submenu-container'><div id='submenu'><a href='/view_trips'><button class='submenu-button'>View Trips</button></a>"
             + "<a href='/account_settings'><button class='submenu-button'>Account Settings</button></a>"
             + "<button class='submenu-button' onclick='logOut();'>Log Out</button></div></div>";
     }
     else {
-        navBar.innerHTML = "<a href='/'><img src='/static/images/banner.png' alt='Team 1 Travel Planner banner'"
-            + "width='961' height='51'></a> <a href='/sign_up'><button class='button signup'>SIGN UP</button></a> "
+        navBar.innerHTML = "<a href='/static'><img src='/static/images/banner.png' alt='Team 1 Travel Planner banner'"
+            + " width='961' height='51'></a> <a href='/sign_up'><button class='button signup'>SIGN UP</button></a> "
             + "<button class='button login' id='login-button' onclick='toggleLoginMenu();'>LOGIN</button>"
             + "<div id='loginmenu-container'><div id='loginmenu'><div class='plan-form'><form id='tripForm'><table id='logintable'>"
             + "        <tr><td>User ID:</td><td><input type='text' id='userID' required></td></tr>"
@@ -87,9 +88,9 @@ function logOut() {
     if (currentUrl.includes("account_settings")||currentUrl.includes("view_trips")){
         location.href= "/";
     }
-}
+
 //simple log in action. temporary
-function logIn() {
-    userName = "username";
-    displayNavBar();
+// function logIn() {
+//     userName = "username";
+//     displayNavBar();
 }
