+++ app/templates/index.html
@@ -4,7 +4,6 @@
 
 
 {% block pagebody %}
-<div id="indexcontent"><a href="/plan_trip"><button class="button plan"><img src="/images/plan_icon.png"
+<div id="indexcontent"><a href="/plan_trip"><button class="button plan"><img src="/static/images/plan_icon.png"
     width="35" height="35" style="vertical-align:middle" alt="Plan Icon"> Plan a Trip</button></a></div>
 {% endblock %}
-
