+++ app/trip/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+blueprint = Blueprint('trip', __name__)
+
+from app.trip import routes
